* Fixes https://github.com/pytorch/pytorch/issues/118675
* Replaces https://github.com/pytorch/pytorch/pull/118994

cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx chenyang78 kadeng chauhang amjames

[ghstack-poisoned]

+++ torch/csrc/distributed/c10d/ParamCommsUtils.hpp
@@ -16,8 +16,8 @@ class TORCH_API ParamCommsDebugInfo : public c10::DebugInfoBase {
       std::tuple<std::string, std::string> pgName,
       int rank,
       std::string&& collName,
-      int64_t inNelems,
-      int64_t outNelems,
+      int inNelems,
+      int outNelems,
       at::ScalarType dType,
       std::vector<int64_t> inSplitSizes,
       std::vector<int64_t> outSplitSizes,
@@ -55,11 +55,11 @@ class TORCH_API ParamCommsDebugInfo : public c10::DebugInfoBase {
     return collectiveName_;
   }
 
-  int64_t getInMessageNelems() const {
+  int getInMessageNelems() const {
     return inMessageNelems_;
   }
 
-  int64_t getOutMessageNelems() const {
+  int getOutMessageNelems() const {
     return outMessageNelems_;
   }
 
@@ -84,8 +84,8 @@ class TORCH_API ParamCommsDebugInfo : public c10::DebugInfoBase {
   int rank_{};
   int worldSize_{};
   std::string collectiveName_;
-  int64_t inMessageNelems_{};
-  int64_t outMessageNelems_{};
+  int inMessageNelems_{};
+  int outMessageNelems_{};
   at::ScalarType dType_ = at::kByte;
   std::vector<int64_t> inputSplitSizes_;
   std::vector<int64_t> outputSplitSizes_;
